Compacted tests
