+++ xmipp3/tests/test_protocol_deep_center_assignment.py
@@ -33,42 +33,69 @@
 # of Neural Networks, specially regarding their training.
 
 # Scipion em imports
-from pyworkflow.tests import BaseTest, DataSet, setupTestProject
 from pwem.protocols import ProtImportParticles, ProtSubSet, exists
+from pyworkflow.tests import BaseTest, DataSet, setupTestProject
 
 # Plugin imports
-from ..protocols import XmippProtDeepCenter
+from ..protocols import XmippProtDeepGlobalAssignment, XmippProtDeepCenter
+
+# ---------------------------- COMMON FUNCTIONS & VARIABLES ----------------------------
+genericError = "There was a problem while running this protocol."
+
+def runImportParticles(cls):
+    """ Import Particles. """
+    args = {'importFrom': ProtImportParticles.IMPORT_FROM_SCIPION,
+            'sqliteFile': cls.particles,
+            'amplitudConstrast': 0.1,
+            'sphericalAberration': 2.0,
+            'voltage': 200,
+            'samplingRate': 0.99,
+            'haveDataBeenPhaseFlipped': True
+            }
+
+    protImport = cls.newProtocol(ProtImportParticles, **args)
+    protImport.setObjLabel('import particles')
+    cls.launchProtocol(protImport)
+    return protImport
+
+def setUpClassBase(cls):
+    setupTestProject(cls)
+
+    # Data
+    cls.dataset = DataSet.getDataSet('10010')
+    cls.particles = cls.dataset.getFile('particles')
+    cls.protImportParts = runImportParticles(cls)
+
+class TestDeepGlobalAssignment(BaseTest):
+    @classmethod
+    def setUpClass(cls):
+        setUpClassBase(cls)
+
+    def test(self):
+        # Creating subset of particles (same for train and test)
+        subset = self.newProtocol(ProtSubSet,
+                                  inputFullSet=self.protImportParts.outputParticles,
+                                  chooseAtRandom=True,
+                                  nElements=400)
+        self.launchProtocol(subset)
+
+        deepGA = self.newProtocol(XmippProtDeepGlobalAssignment,
+                                  inputImageSet=subset.outputParticles,
+                                  inputTrainSet=subset.outputParticles,
+                                  Xdim=128,
+                                  numModels=5,
+                                  numEpochs=1,
+                                  batchSize=32,
+                                  learningRate=0.001,
+                                  sigma=8,
+                                  patience=5)
+        self.launchProtocol(deepGA)
+        self.assertIsNotNone(deepGA.outputParticles, genericError)
 
 class TestDeepCenter(BaseTest):
     @classmethod
-    def runImportParticles(cls):
-        """ Import Particles.
-        """
-        args = {'importFrom': ProtImportParticles.IMPORT_FROM_SCIPION,
-                'sqliteFile': cls.particles,
-                'amplitudConstrast': 0.1,
-                'sphericalAberration': 2.0,
-                'voltage': 200,
-                'samplingRate': 0.99,
-                'haveDataBeenPhaseFlipped': True
-                }
-
-        # Id's should be set increasing from 1 if ### is not in the
-        # pattern
-        protImport = cls.newProtocol(ProtImportParticles, **args)
-        protImport.setObjLabel('import particles')
-        cls.launchProtocol(protImport)
-        return protImport
-
-    @classmethod
     def setUpClass(cls):
-        setupTestProject(cls)
-
-        # Data
-        cls.dataset = DataSet.getDataSet('10010')
-        cls.particles = cls.dataset.getFile('particles')
-
-        cls.protImportParts = cls.runImportParticles()
+        setUpClassBase(cls)
 
     def test(self):
         subset = self.newProtocol(ProtSubSet,
@@ -95,11 +122,11 @@ class TestDeepCenter(BaseTest):
         self.assertTrue(exists(fnModel0), fnModel0 + " does not exist")
         self.assertTrue(exists(fnModel2), fnModel2 + " does not exist")
 
-        self.assertIsNotNone(deepCenter.outputParticles, "There was a problem with Deep Center Predict")
+        self.assertIsNotNone(deepCenter.outputParticles, genericError)
 
         deepCenter2 = self.newProtocol(XmippProtDeepCenter,
                                        inputImageSet=subset.outputParticles,
                                        trainModels=False)
         self.launchProtocol(deepCenter2)
 
-        self.assertIsNotNone(deepCenter2.outputParticles, "There was a problem with Deep Center Predict")
+        self.assertIsNotNone(deepCenter2.outputParticles, genericError)
